Добавил реализацию Лейбл репозиторий

+++ src/main/java/ferros/reository/LabelRepository.java
@@ -3,5 +3,5 @@ package ferros.reository;
 import ferros.model.Label;
 import ferros.model.Post;
 
-public interface LabelRepository extends GenericRepository<Label, Long>{
+public interface LabelRepository extends GenericRepository<Label, Integer>{
 }
